Refactorings: whisper transport removed
