+++ VERSION
@@ -1,1 +1,1 @@
-0.76.5
+0.76.6
