+++ protocol/transport/envelopes_monitor_test.go
@@ -1,4 +1,4 @@
-package whisper
+package transport
 
 import (
 	"testing"
@@ -11,7 +11,6 @@ import (
 
 	"github.com/status-im/status-go/eth-node/crypto"
 	"github.com/status-im/status-go/eth-node/types"
-	"github.com/status-im/status-go/protocol/transport"
 )
 
 var (
@@ -32,7 +31,7 @@ func TestEnvelopesMonitorSuite(t *testing.T) {
 func (s *EnvelopesMonitorSuite) SetupTest() {
 	s.monitor = NewEnvelopesMonitor(
 		nil,
-		transport.EnvelopesMonitorConfig{
+		EnvelopesMonitorConfig{
 			EnvelopeEventsHandler:          nil,
 			MaxAttempts:                    0,
 			MailserverConfirmationsEnabled: false,
