+++ protocol/transport/filters_manager.go
@@ -71,59 +71,59 @@ func NewFiltersManager(persistence KeysPersistence, service FiltersService, priv
 	}, nil
 }
 
-func (s *FiltersManager) Init(
+func (f *FiltersManager) Init(
 	chatIDs []string,
 	publicKeys []*ecdsa.PublicKey,
 ) ([]*Filter, error) {
 
 	// Load our contact code.
-	_, err := s.LoadContactCode(&s.privateKey.PublicKey)
+	_, err := f.LoadContactCode(&f.privateKey.PublicKey)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to load contact code")
 	}
 
 	// Load partitioned topic.
-	_, err = s.loadMyPartitioned()
+	_, err = f.loadMyPartitioned()
 	if err != nil {
 		return nil, err
 	}
 
 	// Add discovery topic.
-	_, err = s.LoadDiscovery()
+	_, err = f.LoadDiscovery()
 	if err != nil {
 		return nil, err
 	}
 
 	// Add public, one-to-one and negotiated filters.
 	for _, chatID := range chatIDs {
-		_, err := s.LoadPublic(chatID)
+		_, err := f.LoadPublic(chatID)
 		if err != nil {
 			return nil, err
 		}
 	}
 
 	for _, publicKey := range publicKeys {
-		_, err := s.LoadContactCode(publicKey)
+		_, err := f.LoadContactCode(publicKey)
 		if err != nil {
 			return nil, err
 		}
 	}
 
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
 	var allFilters []*Filter
-	for _, f := range s.filters {
+	for _, f := range f.filters {
 		allFilters = append(allFilters, f)
 	}
 	return allFilters, nil
 }
 
-func (s *FiltersManager) InitPublicFilters(chatIDs []string) ([]*Filter, error) {
+func (f *FiltersManager) InitPublicFilters(chatIDs []string) ([]*Filter, error) {
 	var filters []*Filter
 	// Add public, one-to-one and negotiated filters.
 	for _, chatID := range chatIDs {
-		f, err := s.LoadPublic(chatID)
+		f, err := f.LoadPublic(chatID)
 		if err != nil {
 			return nil, err
 		}
@@ -132,15 +132,15 @@ func (s *FiltersManager) InitPublicFilters(chatIDs []string) ([]*Filter, error) 
 	return filters, nil
 }
 
-func (s *FiltersManager) InitCommunityFilters(pks []*ecdsa.PrivateKey) ([]*Filter, error) {
+func (f *FiltersManager) InitCommunityFilters(pks []*ecdsa.PrivateKey) ([]*Filter, error) {
 	var filters []*Filter
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
 	for _, pk := range pks {
 
 		identityStr := PublicKeyToStr(&pk.PublicKey)
-		rawFilter, err := s.addAsymmetric(identityStr, pk, true)
+		rawFilter, err := f.addAsymmetric(identityStr, pk, true)
 		if err != nil {
 			return nil, err
 
@@ -155,7 +155,7 @@ func (s *FiltersManager) InitCommunityFilters(pks []*ecdsa.PrivateKey) ([]*Filte
 			OneToOne: true,
 		}
 
-		s.filters[filterID] = filter
+		f.filters[filterID] = filter
 
 		filters = append(filters, filter)
 	}
@@ -163,7 +163,7 @@ func (s *FiltersManager) InitCommunityFilters(pks []*ecdsa.PrivateKey) ([]*Filte
 }
 
 // DEPRECATED
-func (s *FiltersManager) InitWithFilters(filters []*Filter) ([]*Filter, error) {
+func (f *FiltersManager) InitWithFilters(filters []*Filter) ([]*Filter, error) {
 	var (
 		chatIDs    []string
 		publicKeys []*ecdsa.PublicKey
@@ -181,43 +181,43 @@ func (s *FiltersManager) InitWithFilters(filters []*Filter) ([]*Filter, error) {
 		}
 	}
 
-	return s.Init(chatIDs, publicKeys)
+	return f.Init(chatIDs, publicKeys)
 }
 
-func (s *FiltersManager) Reset() error {
+func (f *FiltersManager) Reset() error {
 	var filters []*Filter
 
-	s.mutex.Lock()
-	for _, f := range s.filters {
+	f.mutex.Lock()
+	for _, f := range f.filters {
 		filters = append(filters, f)
 	}
-	s.mutex.Unlock()
+	f.mutex.Unlock()
 
-	return s.Remove(filters...)
+	return f.Remove(filters...)
 }
 
-func (s *FiltersManager) Filters() (result []*Filter) {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+func (f *FiltersManager) Filters() (result []*Filter) {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
-	for _, f := range s.filters {
+	for _, f := range f.filters {
 		result = append(result, f)
 	}
 
 	return
 }
 
-func (s *FiltersManager) Filter(chatID string) *Filter {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-	return s.filters[chatID]
+func (f *FiltersManager) Filter(chatID string) *Filter {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
+	return f.filters[chatID]
 }
 
 // FilterByFilterID returns a Filter with a given Whisper filter ID.
-func (s *FiltersManager) FilterByFilterID(filterID string) *Filter {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-	for _, f := range s.filters {
+func (f *FiltersManager) FilterByFilterID(filterID string) *Filter {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
+	for _, f := range f.filters {
 		if f.FilterID == filterID {
 			return f
 		}
@@ -226,20 +226,19 @@ func (s *FiltersManager) FilterByFilterID(filterID string) *Filter {
 }
 
 // FilterByChatID returns a Filter for given chat id
-func (s *FiltersManager) FilterByChatID(chatID string) *Filter {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+func (f *FiltersManager) FilterByChatID(chatID string) *Filter {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
-	return s.filters[chatID]
+	return f.filters[chatID]
 }
 
-func (s *FiltersManager) FiltersByPublicKey(publicKey *ecdsa.PublicKey) (result []*Filter) {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
+func (f *FiltersManager) FiltersByPublicKey(publicKey *ecdsa.PublicKey) (result []*Filter) {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 	identityStr := PublicKeyToStr(publicKey)
 
-	for _, f := range s.filters {
+	for _, f := range f.filters {
 		if f.Identity == identityStr {
 			result = append(result, f)
 		}
@@ -249,27 +248,27 @@ func (s *FiltersManager) FiltersByPublicKey(publicKey *ecdsa.PublicKey) (result 
 }
 
 // Remove remove all the filters associated with a chat/identity
-func (s *FiltersManager) Remove(filters ...*Filter) error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+func (f *FiltersManager) Remove(filters ...*Filter) error {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
-	for _, f := range filters {
-		if err := s.service.Unsubscribe(f.FilterID); err != nil {
+	for _, filter := range filters {
+		if err := f.service.Unsubscribe(filter.FilterID); err != nil {
 			return err
 		}
-		if f.SymKeyID != "" {
-			s.service.DeleteSymKey(f.SymKeyID)
+		if filter.SymKeyID != "" {
+			f.service.DeleteSymKey(filter.SymKeyID)
 		}
-		delete(s.filters, f.ChatID)
+		delete(f.filters, filter.ChatID)
 	}
 
 	return nil
 }
 
 // Remove remove all the filters associated with a chat/identity
-func (s *FiltersManager) RemoveNoListenFilters() error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+func (f *FiltersManager) RemoveNoListenFilters() error {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 	var filterIDs []string
 	var filters []*Filter
 
@@ -279,31 +278,31 @@ func (s *FiltersManager) RemoveNoListenFilters() error {
 			filters = append(filters, f)
 		}
 	}
-	if err := s.service.UnsubscribeMany(filterIDs); err != nil {
+	if err := f.service.UnsubscribeMany(filterIDs); err != nil {
 		return err
 	}
 
 	for _, filter := range filters {
 		if filter.SymKeyID != "" {
-			s.service.DeleteSymKey(filter.SymKeyID)
+			f.service.DeleteSymKey(filter.SymKeyID)
 		}
-		delete(s.filters, filter.ChatID)
+		delete(f.filters, filter.ChatID)
 	}
 
 	return nil
 }
 
 // Remove remove all the filters associated with a chat/identity
-func (s *FiltersManager) RemoveFilterByChatID(chatID string) (*Filter, error) {
-	s.mutex.Lock()
-	filter, ok := s.filters[chatID]
-	s.mutex.Unlock()
+func (f *FiltersManager) RemoveFilterByChatID(chatID string) (*Filter, error) {
+	f.mutex.Lock()
+	filter, ok := f.filters[chatID]
+	f.mutex.Unlock()
 
 	if !ok {
 		return nil, nil
 	}
 
-	err := s.Remove(filter)
+	err := f.Remove(filter)
 	if err != nil {
 		return nil, err
 	}
@@ -312,28 +311,28 @@ func (s *FiltersManager) RemoveFilterByChatID(chatID string) (*Filter, error) {
 }
 
 // LoadPartitioned creates a filter for a partitioned topic.
-func (s *FiltersManager) LoadPartitioned(publicKey *ecdsa.PublicKey, identity *ecdsa.PrivateKey, listen bool) (*Filter, error) {
-	return s.loadPartitioned(publicKey, identity, listen, false)
+func (f *FiltersManager) LoadPartitioned(publicKey *ecdsa.PublicKey, identity *ecdsa.PrivateKey, listen bool) (*Filter, error) {
+	return f.loadPartitioned(publicKey, identity, listen, false)
 }
 
 // LoadEphemeral creates a filter for a partitioned/personal topic.
-func (s *FiltersManager) LoadEphemeral(publicKey *ecdsa.PublicKey, identity *ecdsa.PrivateKey, listen bool) (*Filter, error) {
-	return s.loadPartitioned(publicKey, identity, listen, true)
+func (f *FiltersManager) LoadEphemeral(publicKey *ecdsa.PublicKey, identity *ecdsa.PrivateKey, listen bool) (*Filter, error) {
+	return f.loadPartitioned(publicKey, identity, listen, true)
 }
 
 // LoadPersonal creates a filter for a personal topic.
-func (s *FiltersManager) LoadPersonal(publicKey *ecdsa.PublicKey, identity *ecdsa.PrivateKey, listen bool) (*Filter, error) {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+func (f *FiltersManager) LoadPersonal(publicKey *ecdsa.PublicKey, identity *ecdsa.PrivateKey, listen bool) (*Filter, error) {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
 	chatID := PersonalDiscoveryTopic(publicKey)
-	if _, ok := s.filters[chatID]; ok {
-		return s.filters[chatID], nil
+	if _, ok := f.filters[chatID]; ok {
+		return f.filters[chatID], nil
 	}
 
 	// We set up a filter so we can publish,
 	// but we discard envelopes if listen is false.
-	filter, err := s.addAsymmetric(chatID, identity, listen)
+	filter, err := f.addAsymmetric(chatID, identity, listen)
 	if err != nil {
 		return nil, err
 	}
@@ -347,28 +346,28 @@ func (s *FiltersManager) LoadPersonal(publicKey *ecdsa.PublicKey, identity *ecds
 		OneToOne: true,
 	}
 
-	s.filters[chatID] = chat
+	f.filters[chatID] = chat
 
 	return chat, nil
 
 }
 
-func (s *FiltersManager) loadMyPartitioned() (*Filter, error) {
-	return s.loadPartitioned(&s.privateKey.PublicKey, s.privateKey, true, false)
+func (f *FiltersManager) loadMyPartitioned() (*Filter, error) {
+	return f.loadPartitioned(&f.privateKey.PublicKey, f.privateKey, true, false)
 }
 
-func (s *FiltersManager) loadPartitioned(publicKey *ecdsa.PublicKey, identity *ecdsa.PrivateKey, listen, ephemeral bool) (*Filter, error) {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+func (f *FiltersManager) loadPartitioned(publicKey *ecdsa.PublicKey, identity *ecdsa.PrivateKey, listen, ephemeral bool) (*Filter, error) {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
 	chatID := PartitionedTopic(publicKey)
-	if _, ok := s.filters[chatID]; ok {
-		return s.filters[chatID], nil
+	if _, ok := f.filters[chatID]; ok {
+		return f.filters[chatID], nil
 	}
 
 	// We set up a filter so we can publish,
 	// but we discard envelopes if listen is false.
-	filter, err := s.addAsymmetric(chatID, identity, listen)
+	filter, err := f.addAsymmetric(chatID, identity, listen)
 	if err != nil {
 		return nil, err
 	}
@@ -383,24 +382,24 @@ func (s *FiltersManager) loadPartitioned(publicKey *ecdsa.PublicKey, identity *e
 		OneToOne:  true,
 	}
 
-	s.filters[chatID] = chat
+	f.filters[chatID] = chat
 
 	return chat, nil
 }
 
 // LoadNegotiated loads a negotiated secret as a filter.
-func (s *FiltersManager) LoadNegotiated(secret types.NegotiatedSecret) (*Filter, error) {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+func (f *FiltersManager) LoadNegotiated(secret types.NegotiatedSecret) (*Filter, error) {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
 	chatID := NegotiatedTopic(secret.PublicKey)
 
-	if _, ok := s.filters[chatID]; ok {
-		return s.filters[chatID], nil
+	if _, ok := f.filters[chatID]; ok {
+		return f.filters[chatID], nil
 	}
 
 	keyString := hex.EncodeToString(secret.Key)
-	filter, err := s.addSymmetric(keyString)
+	filter, err := f.addSymmetric(keyString)
 	if err != nil {
 		return nil, err
 	}
@@ -416,25 +415,25 @@ func (s *FiltersManager) LoadNegotiated(secret types.NegotiatedSecret) (*Filter,
 		OneToOne:   true,
 	}
 
-	s.filters[chat.ChatID] = chat
+	f.filters[chat.ChatID] = chat
 
 	return chat, nil
 }
 
 // LoadDiscovery adds 1 discovery filter
 // for the personal discovery topic.
-func (s *FiltersManager) LoadDiscovery() ([]*Filter, error) {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+func (f *FiltersManager) LoadDiscovery() ([]*Filter, error) {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
-	personalDiscoveryTopic := PersonalDiscoveryTopic(&s.privateKey.PublicKey)
+	personalDiscoveryTopic := PersonalDiscoveryTopic(&f.privateKey.PublicKey)
 
 	// Check if filters are already loaded.
 	var result []*Filter
 
 	expectedTopicCount := 1
 
-	if chat, ok := s.filters[personalDiscoveryTopic]; ok {
+	if chat, ok := f.filters[personalDiscoveryTopic]; ok {
 		result = append(result, chat)
 	}
 
@@ -442,7 +441,7 @@ func (s *FiltersManager) LoadDiscovery() ([]*Filter, error) {
 		return result, nil
 	}
 
-	identityStr := PublicKeyToStr(&s.privateKey.PublicKey)
+	identityStr := PublicKeyToStr(&f.privateKey.PublicKey)
 
 	// Load personal discovery
 	personalDiscoveryChat := &Filter{
@@ -453,7 +452,7 @@ func (s *FiltersManager) LoadDiscovery() ([]*Filter, error) {
 		OneToOne:  true,
 	}
 
-	discoveryResponse, err := s.addAsymmetric(personalDiscoveryChat.ChatID, s.privateKey, true)
+	discoveryResponse, err := f.addAsymmetric(personalDiscoveryChat.ChatID, f.privateKey, true)
 	if err != nil {
 		return nil, err
 	}
@@ -461,21 +460,21 @@ func (s *FiltersManager) LoadDiscovery() ([]*Filter, error) {
 	personalDiscoveryChat.Topic = discoveryResponse.Topic
 	personalDiscoveryChat.FilterID = discoveryResponse.FilterID
 
-	s.filters[personalDiscoveryChat.ChatID] = personalDiscoveryChat
+	f.filters[personalDiscoveryChat.ChatID] = personalDiscoveryChat
 
 	return []*Filter{personalDiscoveryChat}, nil
 }
 
 // LoadPublic adds a filter for a public chat.
-func (s *FiltersManager) LoadPublic(chatID string) (*Filter, error) {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+func (f *FiltersManager) LoadPublic(chatID string) (*Filter, error) {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
-	if chat, ok := s.filters[chatID]; ok {
+	if chat, ok := f.filters[chatID]; ok {
 		return chat, nil
 	}
 
-	filterAndTopic, err := s.addSymmetric(chatID)
+	filterAndTopic, err := f.addSymmetric(chatID)
 	if err != nil {
 		return nil, err
 	}
@@ -489,23 +488,23 @@ func (s *FiltersManager) LoadPublic(chatID string) (*Filter, error) {
 		OneToOne: false,
 	}
 
-	s.filters[chatID] = chat
+	f.filters[chatID] = chat
 
 	return chat, nil
 }
 
 // LoadContactCode creates a filter for the advertise topic for a given public key.
-func (s *FiltersManager) LoadContactCode(pubKey *ecdsa.PublicKey) (*Filter, error) {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+func (f *FiltersManager) LoadContactCode(pubKey *ecdsa.PublicKey) (*Filter, error) {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
 	chatID := ContactCodeTopic(pubKey)
 
-	if _, ok := s.filters[chatID]; ok {
-		return s.filters[chatID], nil
+	if _, ok := f.filters[chatID]; ok {
+		return f.filters[chatID], nil
 	}
 
-	contactCodeFilter, err := s.addSymmetric(chatID)
+	contactCodeFilter, err := f.addSymmetric(chatID)
 	if err != nil {
 		return nil, err
 	}
@@ -519,41 +518,41 @@ func (s *FiltersManager) LoadContactCode(pubKey *ecdsa.PublicKey) (*Filter, erro
 		Listen:   true,
 	}
 
-	s.filters[chatID] = chat
+	f.filters[chatID] = chat
 	return chat, nil
 }
 
 // addSymmetric adds a symmetric key filter
-func (s *FiltersManager) addSymmetric(chatID string) (*RawFilter, error) {
+func (f *FiltersManager) addSymmetric(chatID string) (*RawFilter, error) {
 	var symKeyID string
 	var err error
 
 	topic := ToTopic(chatID)
 	topics := [][]byte{topic}
 
-	symKey, ok := s.keys[chatID]
+	symKey, ok := f.keys[chatID]
 	if ok {
-		symKeyID, err = s.service.AddSymKeyDirect(symKey)
+		symKeyID, err = f.service.AddSymKeyDirect(symKey)
 		if err != nil {
 			return nil, err
 		}
 	} else {
-		symKeyID, err = s.service.AddSymKeyFromPassword(chatID)
+		symKeyID, err = f.service.AddSymKeyFromPassword(chatID)
 		if err != nil {
 			return nil, err
 		}
-		if symKey, err = s.service.GetSymKey(symKeyID); err != nil {
+		if symKey, err = f.service.GetSymKey(symKeyID); err != nil {
 			return nil, err
 		}
-		s.keys[chatID] = symKey
+		f.keys[chatID] = symKey
 
-		err = s.persistence.Add(chatID, symKey)
+		err = f.persistence.Add(chatID, symKey)
 		if err != nil {
 			return nil, err
 		}
 	}
 
-	id, err := s.service.Subscribe(&types.SubscriptionOptions{
+	id, err := f.service.Subscribe(&types.SubscriptionOptions{
 		SymKeyID: symKeyID,
 		PoW:      minPow,
 		Topics:   topics,
@@ -571,7 +570,7 @@ func (s *FiltersManager) addSymmetric(chatID string) (*RawFilter, error) {
 
 // addAsymmetricFilter adds a filter with our private key
 // and set minPow according to the listen parameter.
-func (s *FiltersManager) addAsymmetric(chatID string, identity *ecdsa.PrivateKey, listen bool) (*RawFilter, error) {
+func (f *FiltersManager) addAsymmetric(chatID string, identity *ecdsa.PrivateKey, listen bool) (*RawFilter, error) {
 	var (
 		err error
 		pow = 1.0 // use PoW high enough to discard all messages for the filter
@@ -584,12 +583,12 @@ func (s *FiltersManager) addAsymmetric(chatID string, identity *ecdsa.PrivateKey
 	topic := ToTopic(chatID)
 	topics := [][]byte{topic}
 
-	privateKeyID, err := s.service.AddKeyPair(identity)
+	privateKeyID, err := f.service.AddKeyPair(identity)
 	if err != nil {
 		return nil, err
 	}
 
-	id, err := s.service.Subscribe(&types.SubscriptionOptions{
+	id, err := f.service.Subscribe(&types.SubscriptionOptions{
 		PrivateKeyID: privateKeyID,
 		PoW:          pow,
 		Topics:       topics,
@@ -601,9 +600,9 @@ func (s *FiltersManager) addAsymmetric(chatID string, identity *ecdsa.PrivateKey
 }
 
 // GetNegotiated returns a negotiated chat given an identity
-func (s *FiltersManager) GetNegotiated(identity *ecdsa.PublicKey) *Filter {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
+func (f *FiltersManager) GetNegotiated(identity *ecdsa.PublicKey) *Filter {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
 
-	return s.filters[NegotiatedTopic(identity)]
+	return f.filters[NegotiatedTopic(identity)]
 }
